requirements.txt
groq
PyPDF2
python-dotenv
pymongo
streamlit